matrix: introduce a helper.go file to keep matrix.go size reasonable
